update fragmentation to show weighted vdev average (#2)
Fragmentation was always being reported as 0% (issue #2) regardless of the
actual fragmentation.

This was due to the value generated by the zpool_influxdb for the pool being 0,
whereas vdevs have fragmentation.

So now we weight vdev fragmentations by pool size and sum to give an average:

    dev   size  frag  note
    ---   ----  ----  ----
 eg vdev1  80GB  50%  (contributes 4/5)
    vdev2  20GB  10%  (contributes 1/5)
    pool  100GB

 → overall 50 * 4/5 + 10 * 1/5 = 42% weighted pool fragmentation

Thanks to sdalu for reporting this.